requirements.txt
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.3
python-dotenv==1.0.0
simple-websocket==0.10.0
bcrypt==4.0.1